agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Agent Spec FastAPI backend for CopilotKit template"
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "uvicorn",
    "python-dotenv",
    "ag-ui-agent-spec[adapters-langgraph, adapters-wayflowcore]",
    "sqlalchemy",
    "langgraph<1.0.0",
    "langchain_core<1.0.0"
]

[tool.uv.sources]
ag-ui-agent-spec = { path = "../../ag-ui/integrations/agent-spec/python" }
